Convert Championship to LiveView
* Convert controller tests to liveview tests
* Add ChampionshipLive.Index
* Add ChampionshipLive.Show
* Delete Championship Controller and HTML
